+++ examples/capistrano2/Capfile
@@ -3,11 +3,10 @@
 require 'bugsnag-capistrano'
 
 set :bugsnag_api_key, ENV['BUGSNAG_API_KEY'] || 'YOUR_API_KEY'
-set :bugsnag_endpoint, ENV['BUGSNAG_ENDPOINT'] 
+set :bugsnag_endpoint, ENV['BUGSNAG_ENDPOINT']
 set :bugsnag_env, ENV["BUGSNAG_RELEASE_STAGE"] || 'production'
 set :current_revision, ENV["BUGSNAG_REVISION"]
 set :repo_url, ENV["BUGSNAG_REPOSITORY"]
-set :bugsnag_use_ssl, ENV["BUGSNAG_USE_SSL"] || false
 set :app_version, ENV["BUGSNAG_APP_VERSION"]
 
 task :deploy do
