--- examples/rake/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-gem "rake"
-gem "bugsnag-capistrano", path: "../../"
