--- lib/bugsnag-capistrano/tasks.rb
@@ -1,2 +1,0 @@
-require_relative 'deploy'
-load File.join(File.dirname(File.expand_path(__FILE__)), 'tasks/bugsnag-capistrano.rake')
